proccess lora msg on gw on different thread

+++ code/gw/gw_rx_manager.py
@@ -1,5 +1,6 @@
 import os, sys, inspect, logging, lzma
 from timeit import default_timer as timer
+from Queue import Queue
 import time
 
 # modify PYTHONPATH in order to imprt internal modules from parent directory.
@@ -20,6 +21,7 @@ def pseudo_recieve():
     logging.debug("[{}][{}][Entered function]".format(__name__, inspect.currentframe().f_code.co_name))
     os.chdir(en_gw_bridge_dir)
     rx_msgs = filter(os.path.isfile, os.listdir('.'))
+    logging.debug("[{}][{}] rx msgs is: {}".format(__name__, inspect.currentframe().f_code.co_name, str(rx_msgs)))
     for msg in rx_msgs:
       if "NOT_READY" in msg:
          rx_msgs.remove(msg)
@@ -31,6 +33,10 @@ def pseudo_recieve():
     curr_file = open(rx_msgs[0], 'rb')
     curr_msg = curr_file.read()
     curr_file.close()
+    if "rpi_lora_lte" not in curr_msg:
+        print("[pseudo_receive] pseudo recieve illegal msg" + curr_msg[:60])
+        print("[pseudo_receive] ILLEGAL: " + rx_msgs[0])
+        exit(2)
     os.remove(rx_msgs[0])
     os.chdir(gw_dir)
     return curr_msg, get_id_from_filename(rx_msgs[0])
@@ -40,15 +46,52 @@ def extract_metadata(metadata):
     logging.debug("[{}][{}][Entered function]".format(__name__, inspect.currentframe().f_code.co_name + "metadata is: " + metadata))
     metadata_lst = metadata.split('.csv_')
     filename = metadata_lst[0] + '.csv'
+    # print(metadata_lst)
     first, last, sequence_num = metadata_lst[1].split('_')
     return first=='1', last=='1', int(sequence_num), filename
 
-
-def upload_to_cloud(filepath):
-    logging.debug("[{}][{}][Entered function]".format(__name__, inspect.currentframe().f_code.co_name + "filepath is: " + filepath))
-    send_file_to_aws(filepath)
     
+def handle_msgs(rx_fifo, ignored_lst, compression_mode):
+    queues_dict = {}
+    logging.info("[{}] msg handler awake".format(__name__))
+    while True:
+        if rx_fifo.empty():  # no msgs are waiting - go to sleep
+            logging.info("[{}] rx_fifo is empty, going to sleep for {} sec".format(__name__, gw_sleep_time_in_sec))
+            time.sleep(gw_sleep_time_in_sec)
+        else:
+            msg, source_id = rx_fifo.get() 
+            # print("HANDLER:\n" + msg[:60])
+            first, last, sequence_num, filename = extract_metadata(msg.split('\n')[0])
+            msg_data = "\n".join(msg.split('\n')[1::])
+            logging.info("[{}] msg_metadata: {} {} {} {} id {} \n".format(__name__, filename, str(first), str(last), str(sequence_num), source_id, msg_data))
+            logging.debug("[{}] MsgData: {} \n".format(__name__, msg_data))
+            filepath = "{}/{}_{}".format(gw_queues_dir, filename, source_id)
 
+            # first msg - create new file
+            if first:
+                curr_en_queue = open(filepath, 'wb+')
+                queues_dict[filepath] = curr_en_queue
+            queues_dict[filepath].write(msg_data)
+            # last msg - close file and upload to cloud
+            if last:
+                start = timer() 
+                if compression_mode:
+                    queues_dict[filepath].seek(0)
+                    compressed_data = queues_dict[filepath].read()
+                    data = lzma.decompress(compressed_data)
+                    queues_dict[filepath].seek(0)
+                    queues_dict[filepath].write(data)
+                # logging.info("[{}] GW_QUEUES: {}".format(__name__, queues_dict[filepath].read()))
+                queues_dict[filepath].close()
+                upload_to_cloud(filepath)
+                queues_dict.pop(filepath)
+                ignored_lst.append("{}_{}".format(filename, source_id))
+                end = timer()
+                append_gw_stats(filename + "_" + source_id, end-start, compression_mode) # filename, size, start-time, TTH, compressed
+                if check_success(filename, source_id):
+                    print("Something went wrong... output file is not identical to original file.\n")
+                else:
+                    print(filename + " UPLOAD PERFECTLY MATCH :) ")
 
 
 def append_gw_stats(filename, tth, compression_mode): 
@@ -60,43 +103,15 @@ def append_gw_stats(filename, tth, compression_mode):
     gw_stats.close()
 
 
-def run_rx(ignored_lst, compression_mode):
-    logging.info("[{}] Starting run...]".format(__name__))
-    queues_dict = {}
+def run_rx(rx_fifo):
+    logging.info("[{}] rx manager is awake".format(__name__))
     while True:
         # msg = recieve()
         msg, source_id = pseudo_recieve()
+        # print("RUN_RX Entered 50 chars: {}".format(msg[:50]))
         if msg == "__EMPTY_DIR__":
-            logging.info("[{}][COMPRESSION_MODE:{}] bridge_dir is empty".format(__name__,str(compression_mode)))
-            time.sleep(5)
-            continue
+            logging.info("[{}] bridge_dir is empty".format(__name__))
+            time.sleep(gw_sleep_time_in_sec)
             # break
-        first, last, sequence_num, filename = extract_metadata(msg.split('\n')[0])
-        msg_data = "\n".join(msg.split('\n')[1::])
-        logging.info("[{}] msg_metadata: {} {} {} {} id {} \n".format(__name__, filename, str(first), str(last), str(sequence_num), source_id, msg_data))
-        logging.debug("[{}] MsgData: {} \n".format(__name__, msg_data))
-        filepath = "{}/{}_{}".format(gw_queues_dir, filename, source_id)
-
-        # first msg - create new file
-        if first:
-            curr_en_queue = open(filepath, 'wb+')
-            queues_dict[filepath] = curr_en_queue
-        queues_dict[filepath].write(msg_data)
-        # last msg - close file and upload to cloud
-        if last:
-            start = timer() 
-            if compression_mode:
-                queues_dict[filepath].seek(0)
-                compressed_data = queues_dict[filepath].read()
-                data = lzma.decompress(compressed_data)
-                queues_dict[filepath].seek(0)
-                queues_dict[filepath].write(data)
-            queues_dict[filepath].close()
-            queues_dict.pop(filepath)
-            upload_to_cloud(filepath)
-            ignored_lst.append("{}_{}".format(filename, source_id))
-            end = timer()
-            append_gw_stats(filename + "_" + source_id, end-start, compression_mode) # filename, size, start-time, TTH, compressed
-            # if check_success(filename, source_id):
-               # print("Something went wrong... output file is not identical to original file.\n")
-                
+        else:
+            rx_fifo.put((msg,source_id))
